--- src/main/kotlin/commons/exceptions/fun.kt
@@ -1,10 +1,0 @@
-package commons.exceptions
-
-import java.sql.SQLException
-
-/**
- * Returns an [SQLException] with the message "The connection is closed".
- *
- * @return an [SQLException]
- */
-fun connectionClosedException() = SQLException("The connection is closed")
