--- src/main/kotlin/commons/expansions/PairExtensions.kt
@@ -1,3 +1,0 @@
-package commons.expansions
-
-fun <T, K> Pair<T, K>.asMapEntry() = listOf(this).toMap().entries.first()
