--- src/main/kotlin/commons/query/Ping.kt
@@ -1,8 +1,0 @@
-package commons.query
-
-/**
- * A data class only containing a property named ping.
- *
- * Used to check if the connection to the database is still alive.
- */
-data class Ping(val ping: Byte)
