+++ src/main/kotlin/habitat/racoons/QueryRacoon.kt
@@ -1,17 +1,17 @@
 package habitat.racoons
 
-import commons.casting.castEquivalent
-import commons.expansions.asKClass
-import commons.expansions.getRuntimeGeneric
-import commons.expansions.isMarkedNullable
-import commons.expansions.isNullOrOptional
-import commons.query.QueryProcessing
 import habitat.RacoonManager
 import habitat.configuration.RacoonConfiguration
 import habitat.context.ParameterCasterContext
 import habitat.definition.LazyId
 import habitat.definition.Table
 import habitat.definition.TableName
+import internals.casting.castEquivalent
+import internals.expansions.asKClass
+import internals.expansions.getRuntimeGeneric
+import internals.expansions.isMarkedNullable
+import internals.expansions.isNullOrOptional
+import internals.query.QueryProcessing
 import java.sql.ResultSet
 import java.sql.SQLException
 import kotlin.reflect.KClass
