+++ src/main/kotlin/internals/expansions/PairExtensions.kt
@@ -1,0 +1,3 @@
+package internals.expansions
+
+fun <T, K> Pair<T, K>.asMapEntry() = listOf(this).toMap().entries.first()
