--- src/test/kotlin/commons/model/PropertyExtractorKtTest.kt
@@ -1,22 +1,0 @@
-package commons.model
-
-import org.junit.jupiter.api.Assertions.assertEquals
-import org.junit.jupiter.api.Test
-
-internal class PropertyExtractorKtTest {
-    data class TestClass(val a: Int, val b: String)
-
-    @Test
-    fun getValueKTest() {
-        val test = TestClass(1, "test")
-        assertEquals(1, getValueK(test, "a", TestClass::class))
-        assertEquals("test", getValueK(test, "b", TestClass::class))
-    }
-
-    @Test
-    fun getValueTest() {
-        val test = TestClass(1, "test")
-        assertEquals(1, getValue(test, "a"))
-        assertEquals("test", getValue(test, "b"))
-    }
-}
